Fixed issue with known `double` types treating a floating point value as `float` instead of `double` on assignment.

+++ examples/variadic.fx
@@ -4,7 +4,7 @@ using standard::io::console;
 
 def variadic(...) -> void
 {
-	print(...[0]); print();
+	print(double(...[0])); print();
 	print(...[1]); print();
 	print(...[2]); print();
 	print(...[3]); print();
